Contract: Add events
Add events for any updates to the contract state for the functions
implemented so far.

I figure these are needed, because, while writing the tests earlier, I
learned that I can't access the return values of the solidity functions
from my JavaScript test.  Events are the way to get access to the return
values from JavaScript.

https://ethereum.stackexchange.com/questions/3285/how-to-get-return-values-when-function-with-argument-is-called

'It is not currently possible to return values from functions which modify the blockchain. To receive a return value, you can mark read-only functions as "constant".

For non-constant functions, the only way to "return" information is by using Solidity Events, which coalesce as LOG opcodes in the Ethereum Virtual Machine.'

answered Apr 23, 2016 at 7:47

Signed-off-by: Nathalie C. Chan King Choy <nathalie.ckc.123@gmail.com>