Implement #11, part 1: Sales dist functionality
I realized that sales distributor doesn't need to bin the widgets and
post them for sale.  That can happen automatically as soon as the
results are registered by the tester.

So, the sales distributor's functions would just be to update the mask for
the tests to know what goes in each bin, and to update the unit prices.

For the sake of time savings, I'm just setting initial mask & prices
hard coded in constructor.

Update the tests to use test results that would match bin masks.

Signed-off-by: Nathalie C. Chan King Choy <nathalie.ckc.123@gmail.com>