Contract: Finish #11 - recordWidgetTests now bins
Updated recordWidgetTests to add binned widgets to the appropriate bin
list.  If the widget doesn't match any of the masks, then it's too low
quality to sell (not enough passing tests).

Once in the bin list, the widget can be bought by a customer.

Not yet tested, but does compile.

Signed-off-by: Nathalie C. Chan King Choy <nathalie.ckc.123@gmail.com>