Contract: Clean up vars for 3 bins, SafeMath32
Having separate variables for all 3 bins was annoying. Switch to using
arrays where it is easy.  Designate bin0 as the one with unsellable
widgets so that the indices can just be the bin number, for easy array
access.

Tried compiling and realized that SafeMath only applies to 1 type at a
time.  So, use SafeMath for uint256 and make SafeMath32 for uint32,
which is just a copy of the OpenZeppelin library with 32 subbed in for
256.

Signed-off-by: Nathalie C. Chan King Choy <nathalie.ckc.123@gmail.com>