UI: Create bare bones admin page (start #15)
For now, just copy index.html and delete the stuff I think I don't need.

TO DO:
Create a row of 3 table cells, where each column has a category of admin
functions.

I am thinking, for simplicity, each function might have its own HTML
page and then each of those pages offers the fields to type into and a
button to send off the transaction.

TO DO:
The first page (eventually index.html) would have to detect the Ethreum
address and figure out which role it belongs to, to send the admin to
the admin page if an admin account is the current account loaded.

Signed-off-by: Nathalie C. Chan King Choy <nathalie.ckc.123@gmail.com>