UI: Start #15, form data to JS
Figured out how to make a form field on the HTML page that sends the
data input into the field over to the JS file upon clicking button.

Still need to clean up making the value in the form disappear on the
click, but that can probably be done in the JS somehow, later.

Signed-off-by: Nathalie C. Chan King Choy <nathalie.ckc.123@gmail.com>